[NR] #143 - Use an in memory analytics store when running in standalone mode (#25)

+++ app/interfaces/AnalyticsStore.java
@@ -10,7 +10,7 @@ public interface AnalyticsStore {
 
     CompletableFuture<List<Map<String, Object>>> recentEvents(int limit);
 
-    CompletableFuture<Map<Integer, Integer>> pageVisits();
+    CompletableFuture<Map<Integer, Long>> pageVisits();
 
     CompletableFuture<Boolean> isUp();
 }
